separate pipeline state check from reading data
the DataReady events contain the data to be read but are not
intended to be used to also wait for the pipeline to end.  in
my case, when a powershell script failed, a null exception
was thrown because there was no reason yet for while the pipline
was stopping.  moving that check outside of the event handler
lets us use one set of code to read the output and copy
it to the output window, and another piece of code to wait
for the pipline to end.  while waiting, Application.RunEvents is
called so the UI remains responsive.